requirements.txt
simple-salesforce==1.12.5
python-dotenv==1.0.0
modal==0.71.1
requests==2.31.0
tabulate==0.9.0